size

+++ hash.cpp
@@ -27,7 +27,7 @@ void toHex(size_t x)
 
 size_t hash(std::string str)
 {
-  int sum = 0;
+  size_t sum = 0;
   int size = str.length();
   for (int i = 0; i < size; i++)
   {
